Move fields package outside of editobservation
