+++ gnd/src/main/java/com/google/android/gnd/ui/common/ViewModelModule.java
@@ -20,8 +20,8 @@ import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 import com.google.android.gnd.MainViewModel;
 import com.google.android.gnd.ui.editobservation.EditObservationViewModel;
-import com.google.android.gnd.ui.editobservation.field.FieldViewModel;
-import com.google.android.gnd.ui.editobservation.field.PhotoFieldViewModel;
+import com.google.android.gnd.ui.field.FieldViewModel;
+import com.google.android.gnd.ui.field.PhotoFieldViewModel;
 import com.google.android.gnd.ui.home.HomeScreenViewModel;
 import com.google.android.gnd.ui.home.featuresheet.FeatureSheetViewModel;
 import com.google.android.gnd.ui.home.featuresheet.ObservationListViewModel;
