+++ gnd/src/main/java/com/google/android/gnd/ui/editobservation/EditObservationFragment.java
@@ -36,9 +36,9 @@ import com.google.android.gnd.ui.common.BackPressListener;
 import com.google.android.gnd.ui.common.EphemeralPopups;
 import com.google.android.gnd.ui.common.Navigator;
 import com.google.android.gnd.ui.common.TwoLineToolbar;
-import com.google.android.gnd.ui.editobservation.field.FieldFactory;
-import com.google.android.gnd.ui.editobservation.field.FieldView;
-import com.google.android.gnd.ui.editobservation.field.FieldViewModel;
+import com.google.android.gnd.ui.field.FieldFactory;
+import com.google.android.gnd.ui.field.FieldView;
+import com.google.android.gnd.ui.field.FieldViewModel;
 import java.util.ArrayList;
 import java.util.List;
 import javax.inject.Inject;
