+++ gnd/src/main/java/com/google/android/gnd/ui/field/FieldFactory.java
@@ -1,4 +1,4 @@
-package com.google.android.gnd.ui.editobservation.field;
+package com.google.android.gnd.ui.field;
 
 import androidx.annotation.NonNull;
 import com.google.android.gnd.model.form.Field;
