+++ gnd/src/main/java/com/google/android/gnd/ui/field/FieldView.java
@@ -1,4 +1,4 @@
-package com.google.android.gnd.ui.editobservation.field;
+package com.google.android.gnd.ui.field;
 
 import android.view.LayoutInflater;
 import android.widget.FrameLayout;
