+++ gnd/src/main/java/com/google/android/gnd/ui/field/MultipleChoiceFieldLayout.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.ui.editobservation.field;
+package com.google.android.gnd.ui.field;
 
 import android.content.Context;
 import android.util.AttributeSet;
