+++ gnd/src/main/res/layout/add_photo_bottom_sheet_dialog.xml
@@ -24,7 +24,7 @@
       type="com.google.android.gnd.model.form.Field" />
     <variable
       name="viewModel"
-      type="com.google.android.gnd.ui.editobservation.field.FieldViewModel" />
+      type="com.google.android.gnd.ui.field.FieldViewModel" />
   </data>
 
   <LinearLayout
