+++ gnd/src/main/res/layout/photo_field.xml
@@ -21,7 +21,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.gnd.ui.editobservation.field.PhotoFieldViewModel" />
+      type="com.google.android.gnd.ui.field.PhotoFieldViewModel" />
   </data>
 
   <com.google.android.material.card.MaterialCardView
