+++ gnd/src/main/res/layout/text_input_field.xml
@@ -26,7 +26,7 @@
       type="com.google.android.gnd.model.form.Field" />
     <variable
       name="viewModel"
-      type="com.google.android.gnd.ui.editobservation.field.FieldViewModel" />
+      type="com.google.android.gnd.ui.field.FieldViewModel" />
   </data>
 
   <LinearLayout
